Split up class management in past and future
